! Some columns containing message ID's were not formated as unsigned integers. (Search.php, ManageSearch.php, upgrade_1-1.sql)
git-svn-id: https://devel.simplemachines.org/code/smf/trunk@5309 a96bbafc-d752-43aa-a6db-eed64a5f2cf9

+++ Sources/ManageSearch.php
@@ -412,7 +412,7 @@ function CreateMessageIndex()
 			$smfFunc['db_query']('', "
 				CREATE TABLE {$db_prefix}log_search_words (
 					ID_WORD " . $index_properties[$context['index_settings']['bytes_per_word']]['column_definition'] . " unsigned NOT NULL default '0',
-					id_msg mediumint(8) unsigned NOT NULL default '0',
+					id_msg int(10) unsigned NOT NULL default '0',
 					PRIMARY KEY (ID_WORD, id_msg)
 				) TYPE=MyISAM", __FILE__, __LINE__);
